requirements.txt
pydantic
PyYAML
superagi-tools
instagrapi 
